+++ pom.xml
@@ -136,6 +136,11 @@
 			<artifactId>aws-java-sdk-s3</artifactId>
 			<version>1.12.599</version>
 		</dependency>
+		<dependency>
+			<groupId>software.amazon.awssdk</groupId>
+			<artifactId>ssm</artifactId>
+			<version>2.21.37</version>
+		</dependency>
 	</dependencies>
 
 	<build>
