--- src/main/java/com/api/vetlens/config/S3Config.java
@@ -1,28 +1,0 @@
-package com.api.vetlens.config;
-
-import com.amazonaws.auth.AWSCredentials;
-import com.amazonaws.auth.AWSStaticCredentialsProvider;
-import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class S3Config {
-
-    private final String accessKey =  System.getProperty("AWS_SECRET_KEY");
-
-    private final String accessSecret = System.getProperty("AWS_ACCESS_SECRET");
-
-    private final String region = System.getProperty("AWS_REGION");
-
-    @Bean
-    public AmazonS3 s3Client() {
-        AWSCredentials credentials = new BasicAWSCredentials(accessKey, accessSecret);
-        return AmazonS3ClientBuilder.standard()
-                .withCredentials(new AWSStaticCredentialsProvider(credentials))
-                .withRegion(region).build();
-    }
-
-}
